# openapi/dindin.yaml
...
openapi: '3.1.0'
info:
  title: DinDin API
  version: '1.0.0'
  description: |
    API do DinDin — carteiras, posições, geladeira, proventos, patrimônio,
    carteira sugerida, simulação e cobrança.
servers:
  - url: https://dindin-4e720.web.app
    description: Produção
  - url: http://localhost:5001
    description: Emuladores
security:
  - firebaseIdToken: []
components:
  securitySchemes:
    firebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ID token do Firebase Auth, enviado em toda requisição.
  responses:
    Unauthorized:
      description: Sem ID token válido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Requisição inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        Corpo de erro produzido pelo `asyncHandler` a partir do `HttpError`.
        A mensagem de 4xx é escrita para a tela e vem em pt-BR; a de 5xx não
        é exposta.
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          description: Código de contrato lido pelo frontend, como `SUBSCRIPTION_REQUIRED`.

    # -----------------------------------------------------------------------
    # Enums que vivem em dindin-models (issue #303)
    # -----------------------------------------------------------------------
...
    SubscriptionStatus:
      type: string
      description: 'Estado da assinatura do usuário (#147).'
      enum: [none, trialing, active, past_due, canceled]
...
    SubscriptionPlan:
      type: string
      enum: [basic]
    SubscriptionInterval:
      type: string
      enum: [month, year]
...
    StorePurchaseRequest:
      type: object
      description: 'Compra feita na loja; o backend valida `credential` com a loja antes de conceder o acesso.'
      required: [platform, productId, credential]
      properties:
        platform: { type: string, enum: [apple, google] }
        productId: { type: string }
        credential:
          type: string
          description: 'Recibo da App Store ou `purchaseToken` do Google Play.'
    PublicSubscription:
      type: object
      description: Visão pública da assinatura em `GET /api/me`, sem ids do provedor.
      required: [status, plan, interval, currentPeriodEnd, cancelAtPeriodEnd]
      properties:
        status: { $ref: '#/components/schemas/SubscriptionStatus' }
        plan:
          oneOf:
            - $ref: '#/components/schemas/SubscriptionPlan'
            - type: 'null'
        interval:
          oneOf:
            - $ref: '#/components/schemas/SubscriptionInterval'
            - type: 'null'
        currentPeriodEnd: { type: [string, 'null'], format: date-time }
        cancelAtPeriodEnd: { type: boolean }
...
paths:
...
  /api/billing/store/purchase:
    post:
      summary: Valida a compra na loja e concede a assinatura (também restaura)
      operationId: registerStorePurchase
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StorePurchaseRequest' }
      responses:
        '200':
          description: Assinatura depois da validação do recibo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicSubscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409':
          description: 'Assinatura já ativa (`ALREADY_SUBSCRIBED`) ou recibo de outra conta (`RECEIPT_ALREADY_USED`)'
        '503':
          description: 'Loja não configurada (`STORE_NOT_CONFIGURED`)'
...